Do streaming reads of leaves during iteration
This changes the tree iterator to read leaf data in a streaming manner.
It also will have the effect of disabling node caching for leaf data. I
think that will be a positive change for now since there is currently
some potential for scans of large topics like /images to blow out the
inner node cache. In the future we may want to reintroduce leaf data
caching at the level of MCAP chunks, but that will require either a
custom MCAP reader or some new features in the existing one.