Switch to 24 byte node IDs
This will allow us to accomodate leaf nodes bigger than 4GB, which is
conceivable if a tree is really poorly dimensioned. It will increase the
size of inner nodes by an additional 512 bytes but that will probably be
swamped by the volume of stats.

This change will break any existing data directories.