Fix bug in tree merge algorithm
Prior to this commit, tree merging was not properly avoiding descending
into nodes in the base tree (tree getting merged into) that did not
change in the overlay (tree getting merged). This resulted in massive
write amplification in both the WAL and storage.

This commit addresses that by making the merge accept a "destination"
tree reader. When inner nodes are merged and one of the participating
trees is the destination, the resulting merged inner node will inherit
the destination node's children, and then overwrite any conflicting
child slots.

This commit also turns on zstd compression for writes to the WAL.