Add support for multiple schemas within a single tree
Track statistics for multiple schemas in a single tree, if there are
multiple. Distinguish between the two schemas in statrange output so the
user can figure out how to handle it. Store statistics objects for each
schema present in the inner nodes of the tree.

This will allow the system to support users changing schemas of topics
over time. I considered instead requiring that each tree have a distinct
schema, but this ended up making "schema hashes" more of a user-facing
concept than I think we will want it to be -- though we don't completely
escape it and will need a schema API down the line.