Do not require that queries specify topics
If topics are not supplied, all topics are queried and merged. This is
implemented by adding a new GetLatestByTopic method to the rootmap. I'm
a little concerned this will end up locking us into SQL databases for
the rootmap since it's a bit beyond what a KV store would give us.

This commit also changes the merging setup to initialize all scans
concurrently.