Improve iterator TTFB
Prior to this commit the tree iterator was assembling a list of leaf
node IDs to process during initialization. Queries spanning e.g the full
time range would need to read and parse every inner node in the tree
prior to returning the first message.

Now we only do the required up-front work to get the first message,
which should generally amount to 5 node lookups. With this change TTFB
is subsecond when querying all my available data (all topics, all time)
and is independent of the time interval requested -- though somewhat
dependent still on the topic count and nature.