Store ancestor data on leaf nodes
Prior to this commit, when an insert collided with an existing leaf
page, we merged the old data and new data into the new leaf.

This would have caused a problem when we implement "messages since
version", since there is no way to distinguish the old data from the new
data for exclusion.

Instead, tree leaves need to be reverse linked lists, pointing to a
possible ancestor if one exists. On read, the appropriate set of
ancestors must be merged together. On write, no read of leaf data from
persistent storage should occur.

This commit breaks the data file format, so any databases will need to
be rebuilt.