Add initial support for range deletion, refactor merge and insert methods
Added initial implementation of (+ tests for) DeleteMessagesInRange,
which constructs a partial tree that represents the deletion of
messages in the given range.

Tombstones are placed in the tree at the location of children that are
fully contained within the range. The partial tree, consisting of all
nodes on the path to deletion (excluding the deleted nodes themselves,
which have been tombstoned in their parents' children arrays), is
returned.

Refactored Merge, Insert, and MergeInserts methods to return a memtree.
Insert and Merge no longer take in a *MemTree to write their output
to. An empty memtree is created within these methods and used for
writing.

Cleaned up method docstrings.

Made NewUnexpectedNodeError public.

Still needs to reconcile statistics correctly.