Add log tailing/get messages by version
This adds support for "GetMessagesSinceVersion". It changes the topics
request parameter of the export endpoint to be a map of topic->version,
rather than a list of topics.

In the export response headers, it puts a map that may be used in the
next request if attempting to tail the logs. It is necessary to track
versions for all topics because the fact that writes may have originated
from different nodes (once we have clustering support) means versions
are not comparable across trees.