Fix possible NPE in getMessages
Prior to this commit, if we got a request to merge two topics, and both
had data on the relevant leaf, but only one had data in the actual
requested range, and the one that did not have the data preceded the one
with the data in the listing of roots returned by the rootmap, then we
would end up with a mismatch between indexes in priority queue elements
and the array of iterators from which to draw next elements.

This could either result in an NPE if attempting to draw from an index
that didn't exist, or in drawing next elements from the wrong iterator
-- although I think if that were happening you would be likely if not
guaranteed to hit an NPE within the same request.