Add grammar for dp3 query language
This adds a participle grammar for the dp3 query language. The language
is designed to support easy expression of merged and as-of joined
results.

The main features of the language are,
* comma operator - used to merge topics (tables)
* precedes/succeeds keywords to express as-of joins
* where clauses to support filtering
* limit/offset, which assume a timestamp sort order.

There is no support yet for variable-length array filtering (will
require some extensions to our ros1msg parser) but there is some syntax
envisioned for it.

Here are some examples of queries:

Query a single topic for all time:
```
from robot /diagnostics
```

Query a single topic for a time range:
```
from robot between '2020-01-01' and '2022-01-01' /diagnostics
```

Query a merge of multiple topics
```
from robot /diagnostics, /tf
```

Express an as-of join, selecting for hard brake events when it was
raining.
```
from robot /is_raining precedes /hard_brake
where /is_raining.raining = 't'
limit 1 offset 10
```