Add query plan module
This adds a module for query "planning", which is really just massaging
our raw query AST into a more convenient structure to work with.

The plan module turns a query AST into a tree of logical plan nodes,
which have a close association with our physical operators. The executor
module will then take the plan tree and convert it to an execution tree.