Add query executor
This adds an iterator-style query executor for the dp3 query language.
The executor converts a plan tree to a tree of executor nodes. Each
executor node implements a Next() method that returns a tuple or io.EOF.
Queries are executed by repeatedly invokimg Next() on the root of the
tree.