Push full unresolved where clause to each scan
This pushes the full unresolved where clause to each scan node in the
plan step, and allows the execution step to resolve schema/field
assignment. This is a little more convenient since the execution step
knows about schemas and the planning step does not.

On the other hand there is more work required in this flow on the
execution side. Today the executor doesn't learn about schemas until it
gets its first tuple from a tree. This is very late to learn about a
mismatch. It would be better if we stored the schemas alongside the tree
and kept them in a cache, to enable query validation without descending
to leaf nodes.