Segregate trees in storage by topic/producer
This partitions the tree storage by topic/producer prefixes. The choice
of topic/producer over producer/topic is under the assumption that
potential direct readers will be more likely to be reading everything on
one topic than all topics on one producer.

A substitution string is used for slashes in the prefixes - $()$.
Slashes are very common in ROS topics so the resulting files are a
little strange - though easy to interpret.