Ensure local disk directory store does atomic put
Prior to this commit, the directory storage provider did a write
straight to final storage, which would result in a partial file if
interrupted. Now we write to a tmpfile and do an atomic rename to the
final location.

On initialization of a directory store, the root directory is walked and
any orphaned tmpfiles are deleted.