Store deleted ranges on leaf nodes
This adds some new metadata to the leaf nodes to enable them to
represent data deletions. When a leaf node is functioning as a deletion,
it has an ancestor node ID and a deletion start/end.

During read, the tree iterator is responsible for reading a
(potentially) linked list of leaf nodes and doing appropriate merging
operations. The overhead of this shouldn't be too much in normal cases,
since it will be rare that multiple deletes will intersect a leaf: the
usual deletion pattern will be to clear out full days of data at a time
on e.g 30 day rotation, in which case only a tiny minority of leaf nodes
will have ancestors.

During write, no copying/filtering of existing data is required.
Deletion is very efficient to write.