Merge partial range deletes into tree
This updates the merge procedure to merge partial delete overlays into
the tree. No handling is performed for read yet.

A partial delete overlay is created by the tree's delete routine. For
cases where one of the bounds of the deletion range intersects a leaf
node, we wish to create a linked list pointing back at the previous
data, to function as a mask over that data during read.

In the tree's delete method, the tree is constructed "blindly" as there
is no existing data to reference -- reconciliation with existing data
and proximately timed writes to the WAL occurs during the tree merge.

Consequently, during tree deletion there are no ancestor IDs in hand to
put in the leaf nodes of the overlay, or even knowledge of whether an
ancestor exists at all. For this situation the delete method writes a
leaf with an empty ancestor ID but a populated delete start/end. During
merge, if there is an ancestor in that location, we populate the
ancestor ID before flushing the node to storage. If there is no ancestor
in that location, the node is dropped (as there is nothing to delete).