Return more structure from GetLeafData
Previously our GetLeafData function returned a node ID and a reader. We
now need the ancestor info, so return a full leaf structure (without the
data -- that's still in the reader).

At a future point, the whole GetLeafData function should hopefully be
removed and replaced with usage of Get. The barrier to this today is
that the data field uses a []byte instead of an io.Reader, and readers
dealing with it will actually need an io.ReadCloser (which makes sense
on read but not on write).

So the challenge we have is the write API and the read API for leaves
are a little bit different from each other (unlike for inner nodes), and
our code is not sophisticated enough to represent that cleanly yet.