Merge deleted data on read
This updates the tree iterator to handle leaf nodes with deletion ranges
in their linked lists. The leaf node iterator has a method to get the
"next iterator". In most cases this is just an mcap reader over the next
leaf in the iteration. If there are multiple ancestors to read, or
deletions to handle, we construct a tree of "concat" and "lazy" message
iterators -- one concat iterator for each node in the list -- and
execute a heap-based merge.

The concat iterators are constructed to skip over (without reading)
ranges deleted by chronologically subsequent deletes.