Pull up nested merge joins
With this commit we will pull up/collapse nested merge joins prior to
execution, which was not implemented during the executor work.

Specifically, a query that looks like this:

  [merge
    [scan foo]
    [merge
      [scan bar]
      [scan baz]]]

Will now be rewritten as:

  [merge
    [scan foo]
    [scan bar]
    [scan baz]]

This rewrite would occur on both the right and left sides if both were
merges, however I don't think that's expressible in the grammar today.