Fix bug in merging out of version order
This fixes a bug in the tree merging code. When we merge the conflicted
children of an inner node, we sort them on version so that if one child
corresponds to a delete we can easily discard all previous children.

Separately, the mergeLevel function expects its arguments of nodes and
trees to be in the same order. We were passing the sorted list of nodes
to mergeLevel, without also sorting the trees in the same order. This
caused us to attempt to look up node IDs in the wrong tree, if out of
order versions were supplied to the merge function.

This commit changes the related function to accept a single array of
util.Pair, so that the elements are always in the same order.