Add node IDs to the end of root nodes
This adds the 24 byte node ID to the end of bare root node data files. A
previous commit started doing this to the end of populated data files.
Now that we can rely the last 24 bytes of a data file being a root ID,
clients can get the latest data from the database without consulting the
server or the rootmap. They just need to list the directory, sort to get
the latest file, and read.