Support parenthetical groupings in the grammar
This addresses some rough edges in the grammar through introduction of
parenthetical groupings and adds some structural validation to the
queries.

Previously the meaning of "and" was a little confused in certain
queries, for instance this would parse:

from my-device /foo, /bar where /foo.a = 10 and /bar.b = 20;

The meaning of this is unclear - what does it mean for both "/foo.a" and
"/bar.b" to have these values? Should they have them at the same
nanosecond? That seems not too useful and should probably warrant some
more explicit syntax.

If the same nanosecond isn't implied, then in this context "and" is the same as an
"or".

With this commit we will reject the query above, and make the user
instead write,

from my-device /foo, /bar where /foo.a = 10 or /bar.b = 20;

The where clause of the query is split into separate expression trees
for each table that is referenced, and within each tree we ensure that
all scans and binary expressions reference the same table - i.e that the
tree does not cross tables. If table crossing occurs we error in the
plan step.