Add `desc` keyword to reverse order of results
This commit adds the `desc` keyword to the grammar and the plan.
The keyword is used to reverse the order of the results of a query.

The tree iterator is updated to support traversal in reverse order.
`MergeNode` is also updated to support reverse time-ordered merging of
streams, if the `desc` keyword is present in the query.