Switch to modernc.org/sqlite
This is a pure go sqlite driver transpiled from the official one in C.
This reduces the clean build time on my laptop from 1 minute to 20
seconds. I think there is a minor performance hit associated with this
change, but not likely one we will care about any time soon given the
way we use sqlite.