Return empty MCAP data from executor for empty resultsets
We defer the initialization of the output writer until we see the first
record, to give us the chance to detect a parsing error due to invalid
request filters and return a 400.

Prior to this commit, if the query returned no data we would respond
with an empty 200. Now we do a 200 with an empty MCAP body.