Chomp semicolons off dot command arguments
It is easy to mistakenly add a semicolon to the end of a dot command,
since semicolons are required at the end of queries. This patch will
chomp trailing semicolons off in that situation, to avoid interpreting
an argument as ending in a semicolon and producing a confusing situation
for the user.