Add db truncate command
This commit adds a new command to perform a logical data truncation at
a particular timestamp. This makes data inserted prior to that timestamp
unreachable by read requests. To prevent reads from accessing deleted
files, it is necessary to truncate tables prior to the time when storage
retention policies would delete their objects.

This commit also includes a number of refactorings and cleanups to
prepare for the truncate command. Notably, it removes the memrootmap
and memversionstore implementations, which are no longer needed. The
version store has also been collapsed into the rootmap.