Add explain support
This adds "explain" support to the grammar. When explain is used, the
query is still executed but all nodes are wrapped with instrumentation
and a tree of timings is sent back to the client in an mcap metadata
record. All message/channel/schema data is dropped on the floor.

This also updates the client to render the metadata nicely in that
situation.