Add support for JSON output encoding
This adds support for JSON output encoding in the service, which you can
get by supplying the Accept: application/json header. This will never be
as performant as streaming the binary data, but is probably suitable for
browser usage and will prevent us needing to implement parsing solutions
for all supported encodings in JS. The command line client continues to
transcode client-side.