Add trigram indexes for string equality search acceleration
This adds a new statistic, TrigramSignature, which is a 12-byte bitset
into which all values for each string field are hashed. This gets used
for acceleration of string equality queries.

This mechanism can also be used to accelerate regex queries, by parsing
the query into a bitset operator tree. This patch doesn't go that far
yet.