Support out-of-query limit/offset/skeleton
This adds support for limit and offset in the query request, applied on
top of any limit/offset settings supplied by the requestor. This allows
clients to paginate a user's own potentially paginated query.

It also adds an additional parameter to the query API, "skeleton".  When
the skeleton flag is used, message and channel contents are stripped out
of the response so all is left is a "skeleton" MCAP containing only
timestamps, metadata, and index structures but no data. This can be
compressed and transmitted over the network with much less bandwidth,
allowing for fairly efficient message-level data in cases where the data
itself won't be rendered (e.g histograms of counts).