Fix parser bug with fixed-length arrays >= 64
Prior to this commit we had some code that was reversing a slice of
bytecodes to form a stack. In the bytecodes are some variable-length
integers for lengths. The code for appending varints onto the "stack to
be reversed" was not anticipating the reversal correctly in all cases,
leading to misbehavior when varints were big enough to be multibyte.