Handle duplicate records on ingestion
This adds handling for duplicate records both when they appear within a
single commit, and when they appear in independent commits. Duplicates
are defined as records in the input that have the same hashing channel
(i.e topic, metadata, logical schemas, everything but the channel/schema
IDs), and the same nanosecond timestamp, and the same MCAP sequence
number.

Merging into the tree now reads data from any leaf that overlaps with
the write. Currently, it reads the MCAP messages and extracts
timestamps/sequence numbers from them. This could be optimized in the
future by maintaining a separate compressed index of times/sequences in
an earlier portion of the leaf node, which could be accesed cheaply
separate from the heavier chunk data.

If leaf data already exists in the area being written to, any new data
that is duplicate from that leaf is dropped from the input. Statistics
computation is moved to the WAL merge step from the treemgr Receive step
to allow for this. If leaf data does not exist, any data streams being
merged as part of the commit are deduplicated at that point.

No deduplication occurs on read, since duplicates are assumed to be
stripped at write time.