Move truncate functionality to the grammar
This changes the top-level parsed structure in the query language from a
"Query" to a "Statement". Statements may be either "Query" or
"Truncate". The truncate syntax is the same as the previous dot command
support in the CLI, which this patch removes. Now truncating a table
looks like this:

  truncate my-producer /topic now;

or

  truncate my-producer /topic '2020-01-01';

As before this is a logical rather than physical deletion.