Fix bug in statistics merging with NaNs
Prior to this commit we would produce NaNs if we merged statistics for
two nodes with nonzero message count but zero non-NaN field values. This
would produce a division by zero, resulting in an inner node that would
fail to serialize to JSON. With this change merging two zero-valued
field statistics results in another zero-valued field statistic.