Optimize single leaf merging
This adds an optimization for single leaves written to unoccupied
leaves. For this situation there is no need to merge MCAP files or
decompress anything at the time of merge. Instead we can copy bytes
straight from the WAL to storage, which is more efficient.

To enable this, we must compute statistics earlier in the process. This
commit adds statistics computation to the tree splitting step. In the
fast path this is the only place it occurs, but leaves are being merged,
merged statistics are computed again at that point.