Extend optimization to overlapping inserts
Prior to this commit, we would copy MCAP bytes directly to storage if we
were merging one leaf into an empty destination leaf. Now we will do the
same even if there's a nonempty destination, as long as there are no
overlaps in the timestamp/sequence pairs of the destination and the new
data.

Further, when there is duplication/overlap, we would previously read
MCAP messages out of the destination leaf to compare against. This
commit now takes advantage of the message keys in the header instead of
reading those messages, making the additional overhead of deduplicating
roughly invariant with respect to message size.