Support degenerate range responses in S3 storage implementation
This mitigates an issue encountered working against the digitalocean
spaces API. It appears to send full objects in response to range
requests sometimes, which is its right. The minio client is not giving
sufficient information to the caller to know this has happened, so we
can't do much if the returned data has the wrong offset, but if too much
length is returned we can at least read just the length we know we want.

It's unclear if the wrong-offset case is possible yet. If so this fix
won't be sufficient.