Wrap S3 read seek closer
When using the minio client against digital ocean spaces, the client
returns a read seek closer in response to range calls but seeking
against that reader must be in terms of the original offsets of the
file, not the logical offsets within the range. Furthermore, seek will
reset the logical bounds of minio's readers so that a subsequent call to
ReadAll or similar will read the entire file.

This implements a wrapper type that wraps such a ReadSeekCloser in a
logical offset and length, and translates logical Seek calls to physical
ones so the user can treat it like a normal rsc.