Support multiple producers in one query
This adds support for multiple producers in a single query. They can be
supplied with a comma as a separator. The usefulness of this is
currently limited because the resultset provides no way to trace a
record back to the producer it came from. That will be pretty easy to
accomplish with some channel metadata.

A wildcard merges from all producers. If there are a lot of producers
active simultaneously, this will cause excessive memory usage until we
put some thought into it.